+++ semana19/projeto/src/business/post/getPost.ts
@@ -1,0 +1,26 @@
+
+import { Request, Response } from "express"
+import { selectPost } from "../../data/post/selectPost";
+
+export const getPost = async (req: Request, res: Response) => {
+   try {
+      let message = "Success!"
+
+      const { id } = req.params
+
+      const post = selectPost(id)
+
+      if (!post) {
+         res.statusCode = 404
+         message = "Post not found"
+         throw new Error(message)
+      }
+      res.status(200).send({ message, post })
+
+   } catch (error) {
+      let message = error.sqlMessage || error.message
+      res.statusCode = 400
+
+      res.send({ message })
+   }
+ }
