+++ semana19/projeto/src/business/user/signup.ts
@@ -1,0 +1,40 @@
+
+import { Request, Response } from "express"
+import { insertUser } from "../../data/user/insertUser";
+import { generateToken } from "../../services/authenticator";
+import { hash } from "../../services/hashManager";
+import { generateId } from "../../services/generateId";
+
+export const signup = async (req: Request, res: Response) => {
+    try {
+       let message = "Success!"
+       const { name, email, password } = req.body
+ 
+       if (!name || !email || !password) {
+          res.statusCode = 406
+          message = '"name", "email" and "password" must be provided'
+          throw new Error(message)
+       }
+ 
+       const id: string = generateId()
+ 
+       const cypherPassword = await hash(password);
+ 
+      
+        await insertUser({
+            id,
+            name,
+            email,
+            password: cypherPassword
+         })
+       const token: string = generateToken({ id })
+ 
+       res.status(201).send({ message, token })
+ 
+    } catch (error) {
+       res.statusCode = 400
+       let message = error.sqlMessage || error.message
+ 
+       res.send({ message })
+    }
+ }
