+++ semana19/projeto/src/data/post/insertPost.ts
@@ -1,0 +1,15 @@
+import { connection } from "../../connection";
+import {post} from '../../types/posts'
+export const insertPost = async(
+   {id, photo, description, type, authorId }:any
+) => {
+
+   await connection("labook_posts")
+   .insert({
+      id,
+      photo,
+      description,
+      type,
+      author_id: authorId
+   })
+}
