+++ semana19/projeto/src/data/post/selectPost.ts
@@ -1,0 +1,10 @@
+import { connection } from "../../connection";
+
+export const selectPost = async (
+    id: string
+) => {
+    const queryResult = await connection("labook_users")
+        .select("*")
+        .where({ id })
+    return queryResult[0]
+}
