+++ semana19/projeto/src/data/user/selectUser.ts
@@ -1,0 +1,11 @@
+import { connection } from "../../connection";
+
+export const selectUser = async (
+    id: string
+) => {
+    const queryResult: any = await connection("labook_posts")
+    .select("*")
+    .where({ id })
+
+    return queryResult[0]
+}
