+++ semana19/projeto/src/types/posts.ts
@@ -1,0 +1,13 @@
+export enum POST_TYPES {
+    NORMAL = "normal",
+    EVENT = "event"
+ }
+ export type post = {
+    id: string,
+    photo: string,
+    description: string,
+    type: POST_TYPES,
+    createdAt: Date,
+    authorId: string
+ }
+
