+++ semana19/projeto/src/types/user.ts
@@ -1,0 +1,10 @@
+export type authenticationData = {
+    id: string
+ }
+ 
+ export type user = {
+    id: string,
+    name: string,
+    email: string,
+    password: string
+ }
